add tests, other fixes
